create symtab.c
